Improve codec crate usability
This adds more documentation to the codec and some more examples. Additionally,
we move the hard dependency on `tokio-util` to a feature, allowing the use
outside of tokio without pulling in all of tokio. This also removes the
`FrameError` given it's very limited use and instead returning the bytes
required to build a full frame. While the tokio codec does not take use of this,
 it may be helpful in custom applications.