Expose bytes and byteorder to derive macro to avoid user import
Currently, when you use any of the derive macros from `skrillax-serde`, you'd
need to manually import/use `bytes` and `byteorder`. This is annoying because
it's not something you want to do and also now requires you to depend on these
crates, even though you don't actually use them directly. I didn't quite
understand why this was necessary and just assumed it was some kind of
limitation of the derive macros, even though I knew other crates somehow found a
 way around it. This now provides a solution for my case, by creating an
'internal' module (one that is not publicly described, but technically visible)
which re-expose the two dependencies. This allows the derive macro to reference
the dependencies through that internal module, instead of the 'normal' modules
they belong to. As such, users now longer need to import these themselves, as
they're used through the re-export.