Adjust size of massive header packet to match official size
I have so far assumed the size of a massive header packet is 5 bytes long, first
 byte indicating if it's a header or content, two bytes following packets, two
bytes opcode. Inspecting actual exchange, though, one notices there's a stray
zero byte at the end. It's unclear to me what this byte is for and most
implementations (apparently also the official client?) will just ignore it.
However, since we want to be accurate, we're going to add this empty byte, and
maybe we'll figure out what it is for.