Complete security handshake split
So far, the security implementation was essentially server side only. Thus, if
you wanted to use it on the client, it was not really possible. Instead of
making the security, as it was, capable of also handling client
responsibilities, it made more sense to split it off into the handshake part and
 the actual encryption part. While we're doing that, it also made sense to
actually support everything Silkroad could support. In particular, the flag
provided in the handshake packets was previously mostly ignored. Now, we
actually use it to determine which features we want to establish. While this may
 not actually happen on the official servers, it made sense to support it. This
also implements the long missing checksum and count byte generation, which will
automatically be set up when starting the handshake. It's currently a little
ugly, given we have to re-implement the generation for each frame type, because
the content is always quite a bit different.