Fix creating too large massive frames
While massive frames may not be encrypted, we should never make them larger than
 0x8000 as they will still be treated as encrypted regardless. This is because
the opcode is contained inside the encrypted data and therefor cannot be used to
 determine if the content is encrypted or not. As such, it's exclusive to the
length, which means our huge frames may not be larger than 0x7FFF.

This would've been a problem in our previous iteration, because each massive
container frame was mapped as one full packet structure. In that case, there
wouldn't be much of a use for massive packets, as they couldn't be larger than
a normal frame. Now, because we bundle the content of all massive containers
belonging to a massive packet together, the content may result in >0x7FFF
because it might be spread across multiple container frames.