Do not assert inequality in counter doc
Generally the two bytes should not be the same, but it's not guaranteed. It is
totally possible for them two return the same value. Similarly, `0` may also
occur. Since it's limited to 0-255 range due to the type anyway, there isn't
much we can actually test here. So for now, we remove the condition.