Provide more documentation and make usage more consistent
Some errors were in places were they didn't make sense because we did certain
things on a level where they don't belong. For example, we handled the opcode at
 a stage where we should already have long matched the opcode to a struct. Thus,
 it would usually be either a NOP or always an error. We were also defining the
protocol in the stream crate, which seemed like the wrong place. It will be
provided in another crate to be optionally used if necessary.