Provide crate for simple protocol definition
In the case of an MMORPG, we generally can't predict which packets we're going
to receive at any given moment. There's multiple ways of dealing with this: each
 place can check if the current packet is the kind of packet it expects, which
would mean we'd do a lot of unnecessary work, or instead use some kind of
matching for specific structs to opcodes. In dynamic languages or ones with
runtime reflection, it's easy to build a handler map with opcode -> handler
function. While it is probably possible in rust, it just does not feel right.
Because rust has different primitives for this kind of problem: enums. For small
 enums, there isn't really an issue, but for a game like Silkroad Online, we
have _many_ different packets. At that scale, we need some help to generate the
conversions & derivations. This crate should provide these helpers. Thankfully,
we can also make this easier in the other crates by automatically implementing
the traits for many structs as well.