Add minimal example client that connects to official servers
To validate that the crates work as we'd expect, this provides a very small
client. This client will connect to a gateway server, perform the security
handshake and exchange the identity information. When this client finishes
successfully, we verified the handshake works and that we can also parse
packets. If the handshake didn't happen, the server would simply close on us.