Refactor packet traits to reduce restrictions
Some of these traits were restricted to be `Sized`, which in turn did not allow
them to be used inside a dynamic context. This tries to remove or at least move
these restrictions such that it makes it possible to be used in more dynamic
ways. This should in most cases make the need for an `OutputProtocol`
unnecessary, and it has therefor been removed. We do believe, however, that an
input protocol is still necessary.