Use more unique names in generated proc macro output for serde
Since we use the names of the fields when spreading the contents of a struct /
enum variant, we cannot use them in our generated code, as otherwise it would
shadow them. Similarly, we can't use them when deserializing for temporaries,
like "items". This now prefixes all the elements with `skrillax_serde` such that
 we can essentially be sure they're not shadowing any existing fields and
values.