Do not store checksum table on stack
While linux and osx are totally fine to store this 65kb array on the stack,
windows is not happy. It's also just a total waste to do this so it makes a lot
more sense to use a box and put it on the heap instead. Obviously it would be
able to build the expanded table as a const function, but that's for later.