Allow for stateful parsing for request-response operations
Silkroad frames are quite simple and easy to parse, even stateless. But making
them useful is already somewhat stateful; we need encryption, and the count byte
 also depends on exchanged information only valid for that stream. Most
operations were also stateless and required no more additional information.
However, there appear to be _some_ operations that defy that expectation. Some
are structured differently depending on external factors, such as if an ID is
for an equipment item or not. Others have different structures depending on the
operation that requested them. For example, the group spawn operation content
depends on what the group spawn start operation said it contains. There are ways
 to implement this before this change, but those are generally not easy to do or
 make correct. Thus, is it easier to provide _some_ support here and make the
final implementation bridge the last mile than leaving this to the
implementation entirely. At worst, we're throwing some performance out of the
window if it stays unused. That's an acceptable tradeoff for now, until we have
a better way to deal with that.