Allow enums to have more complex conditions based on their own fields
We previously only allowed optionals to be conditional on values of other,
previous fields. This makes sense if a field is only present when a given value
is set. However, this isn't quite as useful as having a proper enum would be.
This now allows writing an enum using a tag value which is also contained inside
 the enum. While usually, we'd have a problem with deserialization if we left
out the field, now we'd have a problem with serialization. As such, it is
important that we actually store the value in the variant. For now, we even
allow picking a different attribute that isn't the first, but we might limit
this in the future.