Pivot to using a more dynamic approach to handling packets
Previously, there was a focus on static dispatch. This worked, quite well
actually, but the more packets exist, the larger the enums had to become. In
turn, this brought up the protocol types, complicating everything. Given the
number of possible packets and the diversity of them, we have to provide a lot
of support for it; need to support Box and non-Box variants, a way to compose
them, a nice way to extract the individual packets from them, how to pack a
packet into them. Additionally, we currently have no way to support stream
callbacks for typed packets, because we don't even get to see the packet struct
as the type parameters are usually the protocols - not the packet structs.

The alternative approach used now is to base most things off of `Any` and
provide the necessary conversion at the relevant places. This has two main
downsides: performance and the need for registering packets. The former, while
relevant, is not quite as important, as IO could generally be done separately
from a game thread. It's also not like this _destroys_ the performances, it's
just not _as_ performant due to the pointer indirection. To some degree that is
irrelevant given that some packets would most likely be kept on the heap anyway
due to their size. The need for a registry is unfortunate, but not
insurmountable. The handshake can provide an extension that automatically
registers the packets properly. The only bigger problem is that it's now
important how you register packets now. If you accidentally register a packet
for input but meant output, things could break at runtime instead of either
'just working' or breaking at compile time. This is an obvious downgrade, but it
 provides a lot more flexibility, including allowing us to do callbacks for
individual packets now.

There might be more work towards making this more ergonomic, but for a first
working draft, this seems decent enough.