Carry opcode in dynamic packet
We currently only carry the actual packet struct as the content of this
dynamic packet. While this is generally enough, for more - well -
dynamic use cases, being able to access the opcode is crucial. For
example, this would remove the need to try and downcast the content when
in a stream callback, when knowing the opcode is enough. The opcode is
small enough to carry around without having to worry about the size of
this struct, so it should be fine.